+++ index.html
@@ -66,6 +66,6 @@
     <div>Copyright</div>
   </footer>
   
-  <script src="./js/index.js"></script>
+  <script type="module" src="./js/index.js"></script>
 </body>
 </html>
